REF/TST: collect tests by method
