+++ pandas/tests/series/test_period.py
@@ -49,21 +49,6 @@ class TestSeriesPeriod:
         expected = Series([pd.NaT], dtype="period[D]")
         tm.assert_series_equal(result, expected)
 
-    def test_set_none(self):
-        self.series[3] = None
-        assert self.series[3] is pd.NaT
-
-        self.series[3:5] = None
-        assert self.series[4] is pd.NaT
-
-    def test_set_nan(self):
-        # Do we want to allow this?
-        self.series[5] = np.nan
-        assert self.series[5] is pd.NaT
-
-        self.series[5:7] = np.nan
-        assert self.series[6] is pd.NaT
-
     def test_intercept_astype_object(self):
         expected = self.series.astype("object")
